click Magic Maze 3 times & delete doc shall appear
